pyproject.toml
[project]
name = "smallgpt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "torch>=1.10",
  "pyyaml>=5.4",
  "requests>=2.25",
  "transformers>=4.20.0",
  "accelerate>=0.26.0"
]